IPython Console: Move silent_exec_method and handle_exec_method to ShellWidget
This is also going to be needed by HelpWidget, so it's not exclusive to
NamepaceBrowserWidget.
