+++ spyder/widgets/ipythonconsole/namespacebrowser.py
@@ -9,11 +9,6 @@ Widget that handle communications between the IPython Console and
 the Variable Explorer
 """
 
-from __future__ import absolute_import
-
-import ast
-import uuid
-
 from qtpy.QtCore import QEventLoop
 
 from ipykernel.pickleutil import CannedObject
@@ -74,69 +69,6 @@ class NamepaceBrowserWidget(RichJupyterWidget):
         settings = to_text_string(self.namespacebrowser.get_view_settings())
         code = u"get_ipython().kernel.namespace_view_settings = %s" % settings
         self.silent_execute(code)
-
-    def silent_exec_method(self, code):
-        """Silently execute a kernel method and save its reply
-
-        The methods passed here **don't** involve getting the value
-        of a variable but instead replies that can be handled by
-        ast.literal_eval.
-
-        To get a value see `get_value`
-
-        Parameters
-        ----------
-        code : string
-            Code that contains the kernel method as part of its
-            string
-
-        See Also
-        --------
-        handle_exec_method : Method that deals with the reply
-
-        Note
-        ----
-        This is based on the _silent_exec_callback method of
-        RichJupyterWidget. Therefore this is licensed BSD
-        """
-        # Generate uuid, which would be used as an indication of whether or
-        # not the unique request originated from here
-        local_uuid = to_text_string(uuid.uuid1())
-        code = to_text_string(code)
-        msg_id = self.kernel_client.execute('', silent=True,
-                                            user_expressions={ local_uuid:code })
-        self._kernel_methods[local_uuid] = code
-        self._request_info['execute'][msg_id] = self._ExecutionRequest(msg_id,
-                                                          'silent_exec_method')
-
-    def handle_exec_method(self, msg):
-        """
-        Handle data returned by silent executions of kernel methods
-
-        This is based on the _handle_exec_callback of RichJupyterWidget.
-        Therefore this is licensed BSD.
-        """
-        user_exp = msg['content'].get('user_expressions')
-        if not user_exp:
-            return
-        for expression in user_exp:
-            if expression in self._kernel_methods:
-                # Process kernel reply
-                method = self._kernel_methods[expression]
-                reply = user_exp[expression]
-                data = reply.get('data')
-                if 'get_namespace_view' in method:
-                    view = ast.literal_eval(data['text/plain'])
-                    self.sig_namespace_view.emit(view)
-                elif 'get_var_properties' in method:
-                    properties = ast.literal_eval(data['text/plain'])
-                    self.sig_var_properties.emit(properties)
-                elif 'load_data' in method or 'save_namespace' in method:
-                    self._kernel_message = ast.literal_eval(data['text/plain'])
-                    self.sig_error_message.emit()
-
-                # Remove method after being processed
-                self._kernel_methods.pop(expression)
 
     def get_value(self, name):
         """Ask kernel for a value"""
